Created documented sub-package for user mode events and moved system info into it
